fix casts and declarations to fix compilation on my system

+++ auto_nox.c
@@ -2390,7 +2390,7 @@ void auto_start_at_homoclinic()
   if(HomoFlag==2)
     xAuto.iequib=-2;
   flag=get_homo_info(HomoFlag,&xAuto.nunstab,&xAuto.nstab,homo_l,homo_r);
-  if(flag)do_auto(opn,close,Auto.itp);
+  if(flag)do_auto(opn,(int)close,Auto.itp);
 
   
 }
